Fix api::BigInt::fromLong declaration

+++ core/src/api_impl/BigIntImpl.cpp
@@ -121,7 +121,7 @@ namespace ledger { namespace core { namespace api {
                 }
             }
 
-            std::shared_ptr<BigInt> BigInt::fromLong(long long value) {
+            std::shared_ptr<BigInt> BigInt::fromLong(int64_t value) {
                 return std::make_shared<BigIntImpl>(ledger::core::BigInt::fromScalar(value));
             }
